requirements.txt
beautifulsoup4==4.10.0
certifi== 2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
Flask==2.0.2
html5lib==1.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
pip==21.3.1
protobuf==3.19.1
requests==2.26.0
setuptools==58.3.0
six==1.16.0
soupsieve==2.3
urllib3==1.26.7
webencodings==0.5.1
Werkzeug==2.0.2
wheel==0.37.0